+++ src/bin/cmds/publish.rs
@@ -1,6 +1,11 @@
 use super::{args, get};
 use clap::{App, ArgMatches, SubCommand};
-use elba::util::{config::Config, errors::Res};
+use elba::{
+    cli::backend,
+    util::{config::Config, errors::Res},
+};
+use failure::ResultExt;
+use std::env::current_dir;
 
 pub fn cli() -> App<'static, 'static> {
     SubCommand::with_name("publish")
@@ -9,6 +14,20 @@ pub fn cli() -> App<'static, 'static> {
 }
 
 pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
-    let _bck = get::index(c, args);
-    unimplemented!()
+    let project = current_dir().context(format_err!(
+        "couldn't get current dir; doesn't exist or no permissions..."
+    ))?;
+
+    let bck = get::index(c, args)?;
+    let bck_text = bck.to_string();
+    let bcx = get::build_ctx(c, args);
+    let ctx = backend::BackendCtx {
+        index: bck,
+        data_dir: c.directories.data.clone(),
+    };
+
+    // TODO: We require verification of packages before publishing; maybe not?
+    backend::publish(&bcx, &ctx, &project, true)?;
+
+    Ok(format!("published package to index {}", bck_text))
 }