+++ src/bin/cmds/test.rs
@@ -33,17 +33,7 @@ pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
         "couldn't get current dir; doesn't exist or no permissions..."
     ))?;
 
-    let logger = get::logger(c, args);
-
-    let ctx = build::BuildCtx {
-        indices: c.indices.to_owned(),
-        global_cache: c.layout(),
-        logger,
-        threads: get::threads(c, args),
-        shell: c.shell(),
-        offline: args.is_present("offline"),
-        opts: get::idris_opts(c, args),
-    };
+    let ctx = get::build_ctx(c, args);
 
     // This is where our default codegen backend is set
     let backend = get::backends(c, args);